requirements.txt
SpeechRecognition
gTTS
pygame
google-generativeai
PyAudio